% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpcv.R
\name{awkHelper}
\alias{awkHelper}
\title{Read in plantCV csv output in wide or long format}
\usage{
awkHelper(inputFile, filters, awk = NULL)
}
\arguments{
\item{filters}{If a very large pcv output file is read then it may be desireable to subset it before reading it into R, either for ease of use or because of RAM limitations. The filter argument works with "COLUMN in VALUES" syntax. This can either be a character vector or a list of character vectors. In these vectors there needs to be a column name, one of " in ", " is ", or " = ", then a set of comma delimited values to filter that column for (see examples). Note that this and `awk` both use awk through pipe(). This functionality will not work on a windows system.}

\item{awk}{As an alternative to `filters` a direct call to awk can be supplied here, in which case that call will be used through pipe().}

\item{filepath}{Path to csv file of plantCV output.}

\item{mode}{One of "wide" or "long", partial string matching is supported. Priority here is given to reducing dependencies, so long to wide transformation is done using lapply. This can cause problems if there are missing values in `traitCol`}

\item{singleValueOnly}{Logical, should only single value traits be returned? If `mode` = "wide" then this should almost always be set to T.}

\item{traitCol}{Column with phenotype names, defaults to "trait". This should generally not need to be changed from the default.}

\item{labelCol}{Column with phenotype labels (units), defaults to "label". This should generally not need to be changed from the default. This is used with `traitCol` when `mode`="wide" to identify unique traits since some may be ambiguous (ellipseCenter.x vs ellipseCenter.y, etc)}

\item{valueCol}{Column with phenotype values, defaults to "value". This should generally not need to be changed from the default.}

\item{multiValPattern}{If `singleValueOnly`=TRUE then this is used to identify multi value traits. By default this is "hist|frequencies". If this argument has length of 1 then it is taken as either a single phenotype or a regex pattern to find values of `trait` that are multi-value phenotypes. Alternatively this can be a vector of phenotype names to remove (see examples).}

\item{reader}{The function to use to read in data, defaults to "read.csv". Other useful options are "vroom" and "fread", from the vroom and data.table packages, respectively. With files that are still very large after subsetting "fread" or "vroom" should be used.}

\item{...}{Other arguments passed to the reader function. In the case of 'vroom' and 'fread' there are several defaults provided already which can be overwritten with these extra arguments.}
}
\description{
Read in plantCV csv output in wide or long format
}
\keyword{long}
\keyword{pcv,}
\keyword{read.csv,}
\keyword{wide,}
