% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellwetherbox.R
\name{bellwetherBox}
\alias{bellwetherBox}
\title{Boxplots for bellwether data}
\usage{
bellwetherBox(
  df = NULL,
  x = NULL,
  y = NULL,
  time = NULL,
  timeCol = "DAS",
  compare = F,
  fill = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame to use. Should be in wide format similar to output from read.pcv.bw}

\item{x}{X axis variable name}

\item{y}{Y axis variable name}

\item{time}{Value of timeCol to use in subsetting data. If NULL (default) then timeCol is coerced to numeric and the max is used.}

\item{timeCol}{Column of time data to use. Defaults to DAS corresponding to Days after Start from read.pcv.bw}

\item{compare}{Groups to compare. By default this is set to FALSE, which corresponds to no testing. Other values of compare are passed to fixCompare to make t.test comparisons using ggpubr. In short, NULL will run all pairwise T tests, a single value of the X axis variable will compare that level to all other levels of the X variable, alternatively this can be a list as used by ggpubr: list(c("level1", "level2"), c("level1", "level3"))}

\item{fill}{Optional variable to use for boxplot color. If set to T then the x variable is used. If this is set to a different variable than x then care should be taken with compare.}

\item{...}{Additional parameters passed to geom_boxplot}
}
\value{
Returns a ggplot object
}
\description{
Boxplots for bellwether data
}
\examples{
df <- read.pcv.bw("pcvTestData.csv")
bellwetherBox(df, x="genotype", y = "area.pixels", time = 19)
}
\keyword{Bellwether,}
\keyword{ggplot}
