% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwwater.R
\name{bw.water}
\alias{bw.water}
\title{Read in lemnatech watering data from metadata.json files}
\usage{
bw.water(file = NULL, envKey = "environment")
}
\arguments{
\item{file}{Path to a json file of lemnatech metadata.}

\item{envKey}{Character string representing the json key for environment data. By default this is set to "environment". Currently there are no situations where this makes sense to change.}
}
\value{
A data frame containing the bellwether watering data
}
\description{
Read in lemnatech watering data from metadata.json files
}
\examples{
wateringData<-bw.water("example.json") 
}
\keyword{long}
\keyword{pcv,}
\keyword{read.csv,}
\keyword{wide,}
