% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw_vignette_small_emd}
\alias{bw_vignette_small_emd}
\title{pcv.emd output from a subset of bellwether data.}
\format{
A list containing data and plot elements.
}
\usage{
data(bw_vignette_small_emd)
}
\description{
A list containing data and plot elements.
}
\details{
The "data" component is a data.frame with 363,609 rows and 9 columns.

\itemize{
  \item i: The i-th row being compared
  \item j: The j-th row being compared
  \item emd: The Earth Mover's Distance between hue histograms between rows i and j.
  \item fertilizer_i: Fertilizer level for row i
  (these metadata terms were controlled by the \code{reorder} argument in \code{pcv.emd}).
  \item genotype_i: Genotype for row i
  \item  DAS_i: DAS for row i
  \item fertilizer_j : Fertilizer level for row j
  \item genotype_j : Genotype for row j
  \item DAS_j: DAS for row j
}
}
\keyword{datasets}
