% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDraws.R
\name{combineDraws}
\alias{combineDraws}
\title{Combine Draws From brms Models}
\usage{
combineDraws(..., names = NULL, message = TRUE)
}
\arguments{
\item{...}{Some number of brmsfit objects and/or dataframes of draws
(should generally be the same type of model fit to different data)}

\item{names}{Optional vector of names for the models/data.frames. The default (NULL) will use
the object names of arguments to \code{...} as a vector of names to uniquely identify columns in
the output data frame. See details for using this with \code{do.call} on a list of models.}

\item{message}{Logical, should messages about possible problems be printed? Default is TRUE.
This will warn if models may not have converged, if there are different numbers of draws in
the objects, or if models have different formulations.}
}
\value{
A data.frame of posterior draws, labeled to show which object they come from.

Returns a dataframe of posterior draws.
}
\description{
Helper function for binding draws from several \code{brms} models to make a data.frame
for use with \code{brms::hypothesis()}. This will also check that the draws are comparable using
basic model metrics.
}
\details{
If you fit models as part of a loop/apply function and end up with a list of models it may be
helpful to call this function on the list. In that case object names are not parsed well from
the list by default so passing the \code{names} argument is helpful and can be done as
\code{do.call(combineDraws, c(fits, list(names = names(fits))))}.
}
\examples{
# note that this example will fit several models using Stan and may run slowly.
\donttest{
data(fit)
fit_1 <- fit_2 <- fit
x <- combineDraws(fit_1, fit_2, fit)
draws_1 <- as.data.frame(fit_1)
draws_1 <- draws_1[, grepl("^b_", colnames(draws_ef))]
x2 <- combineDraws(fit_2, draws_1)
}

}
\keyword{brms}
