\name{corn}
\alias{corn}
\docType{data}

\title{
Corn data
}

\description{
NIR spectra and moisture of 80 Corn samples.
}

\usage{data(corn)}
\format{
   A list with two matrices, \code{spectra} and \code{moisture}.
}

\details{
   This is a part of Corn dataset, which was downloaded from Eigenvector Research, Inc. website (https://eigenvector.com/resources/data-sets/), where it is availble publicly. This dataset contains several NIR spectra of corn samples recorded using different instruments. For our examples we took "mp5" spectra and corrected them using Standard Normal Variate transformation.
}

\source{
1. Eigenvector Research, Inc. (https://eigenvector.com/resources/data-sets/)
}

\keyword{datasets}
