% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd.R
\name{emd1d}
\alias{emd1d}
\title{Earth Mover's Distance between spectral histograms}
\usage{
emd1d(s1, s2)
}
\arguments{
\item{s1}{Histogram as a numeric vector of counts per position.}

\item{s2}{Histogram as a numeric vector of counts per position. Must be the same length as s1.}
}
\value{
Returns EMD between two samples as a numeric.
}
\description{
emd1d computes 1 dimension EMD for two samples.
}
\examples{

set.seed(123)
s1 <- hist(rnorm(10000, 50, 10), breaks = seq(1, 100, 1))$counts
s2 <- hist(rlnorm(9000, log(30), 0.25), breaks = seq(1, 100, 1))$counts
plot(s2, type = "l"); lines(s1)
emd1d(s1, s2)

}
\keyword{distance,}
\keyword{earth}
\keyword{emd,}
\keyword{histogram}
\keyword{mover's}
\keyword{multi-value}
\keyword{trait,}
