% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fit}
\alias{fit}
\title{Example model}
\format{
A brmsfit object and a growthss object
}
\usage{
fit
}
\description{
A logistic brms model of simulated growth data.
}
\examples{
\donttest{
data(fit, package = "pcvr")
summary(fit)
}

}
\keyword{datasets}
