% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowth}
\alias{fitGrowth}
\title{Ease of use brms wrapper function for fitting growth models specified by \code{growthSS}}
\usage{
fitGrowth(
  ss,
  iter = 2000,
  cores = getOption("mc.cores", 1),
  chains = 4,
  prior = NULL,
  backend = "cmdstanr",
  ...
)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{iter}{A number of iterations to sample for each chain.
By default half this length is taken as warm-up for the MCMC algorithm.
This defaults to 2000.}

\item{cores}{A number of cores to run in parallel.
This defaults to 1 if the "mc.cores" option is not set.
Generally this is specified as one core per chain so that the model is fit in parallel.}

\item{chains}{A number of markov chains to use, this defaults to 4.}

\item{prior}{A \code{brmsprior} object if \code{growthSS} did not have priors specified.
If left NULL (the default) and ss does not contain priors then a warning is
issued but the model will still attempt to fit.}

\item{backend}{A backend for brms to use Stan through.
This defaults to use "cmdstanr".}

\item{...}{Additional arguments passed to \code{brms::brm}.}
}
\value{
A \code{brmsfit} object, see \code{?`brmsfit-class`} for details.
}
\description{
Ease of use brms wrapper function for fitting growth models specified by \code{growthSS}
}
\examples{

## Not run:
simdf<-growthSim("logistic", n=20, t=25,
   params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))
ss<-growthSS(model = "logistic", form=y~time|id/group, sigma="spline",
  df=simdf, priors = list("A"=130, "B"=12, "C"=3))
lapply(ss,class)
ss$initfun()
# fit_test <- fitGrowth(ss, iter = 1000, cores = 2, chains = 2, backend = "cmdstanr",
#               control = list(adapt_delta = 0.999, max_treedepth = 20)) # optional controls
## End(Not run)
}
\keyword{Bayesian,}
\keyword{brms}
