% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthflexsurv}
\alias{fitGrowthflexsurv}
\title{Ease of use wrapper function for fitting growth models specified by \code{growthSS}}
\usage{
fitGrowthflexsurv(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{flexsurv::flexsurvreg}.}
}
\value{
A \code{survreg} object.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{flexsurv}
