% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthlm}
\alias{fitGrowthlm}
\title{Ease of use lm wrapper function for fitting growth models specified by \code{mvSS}}
\usage{
fitGrowthlm(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{mvSS}.}

\item{...}{Additional arguments passed to \code{stats::lm}.}
}
\value{
An \code{lm} object, see \code{?lm} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nls}
