% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthmgcvgam}
\alias{fitGrowthmgcvgam}
\title{Ease of use mgcv wrapper function for fitting gams specified by \code{growthSS}}
\usage{
fitGrowthmgcvgam(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{mgcv::gam}.}
}
\value{
An \code{gam} object, see \code{?gam} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{gam}
\keyword{mgcv,}
