% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNlme}
\alias{fitGrowthNlme}
\title{Ease of use nlme wrapper function for fitting growth models specified by \code{growthSS}}
\usage{
fitGrowthNlme(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{nlme::nlme}.}
}
\value{
An \code{nlme} object, see \code{?nlme} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nlme}
