% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthnlmegam}
\alias{fitGrowthnlmegam}
\title{Ease of use lme wrapper function for fitting gams specified by \code{growthSS}}
\usage{
fitGrowthnlmegam(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{nlme::lmeControl}.}
}
\value{
An \code{lme} object, see \code{?lme} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{gam}
\keyword{nlme}
