% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNlrq}
\alias{fitGrowthNlrq}
\title{Ease of use nlrq wrapper function for fitting growth models specified by \code{growthSS}}
\usage{
fitGrowthNlrq(ss, cores = getOption("mc.cores", 1), ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{cores}{Optionally specify how many cores to run in parallel if ss$taus is >1L. Defaults to 1 if mc.cores option is not set globally.}

\item{...}{Additional arguments passed to \code{nlme::nlme}.}
}
\value{
An \code{nlrqModel} object if tau is length of 1 or a list of such models for multiple taus,
  see \code{?quantreg::nlrq} or \code{?nls::nlsModel} for details.
}
\description{
This will generally be called from \code{fitGrowth}.
}
\keyword{nls}
