% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNlrq_gam}
\alias{fitGrowthNlrq_gam}
\title{Ease of use rq wrapper function for fitting gams specified by \code{growthSS}}
\usage{
fitGrowthNlrq_gam(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{quantreg::rq}.}
}
\value{
An \code{lm} object, see \code{?lm} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nls}
