% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNlrqGam}
\alias{fitGrowthNlrqGam}
\title{Ease of use rq wrapper function for fitting gams specified by \code{growthSS}}
\usage{
fitGrowthNlrqGam(ss, cores = getOption("mc.cores", 1), ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{cores}{number of cores to run in parallel}

\item{...}{Additional arguments passed to \code{quantreg::rq}.}
}
\value{
An \code{lm} object, see \code{?lm} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nls}
