% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNls}
\alias{fitGrowthNls}
\title{Ease of use nls wrapper function for fitting growth models specified by \code{growthSS}}
\usage{
fitGrowthNls(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{stats::nls}.}
}
\value{
An \code{nls} object, see \code{?nls} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nls}
