% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowthNlsGam}
\alias{fitGrowthNlsGam}
\title{Ease of use lm wrapper function for fitting gams specified by \code{growthSS}}
\usage{
fitGrowthNlsGam(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to \code{stats::lm}.}
}
\value{
An \code{lm} object, see \code{?lm} for details.
}
\description{
Helper function generally called from \link{fitGrowth}.
}
\keyword{nls}
