% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvcrossval.R
\name{getCrossvalParams}
\alias{getCrossvalParams}
\title{Returns parameters for cross-validation based on 'cv' value}
\usage{
getCrossvalParams(cv, nobj)
}
\arguments{
\item{cv}{settings for cross-validation provided by user}

\item{nobj}{number of objects in calibration set}
}
\description{
Returns parameters for cross-validation based on 'cv' value
}
