% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxpvorth.R
\name{getxpvorth}
\alias{getxpvorth}
\title{Generates the orthogonal part for Xpv}
\usage{
getxpvorth(q.k, X.k, PRM)
}
\arguments{
\item{q.k}{vector with orthogonal distances for cross-validation set for segment k}

\item{X.k}{matrix with local validation set for segment k}

\item{PRM}{projecton matrix for orthogonalization of residuals}
}
\description{
Generates the orthogonal part for Xpv
}
