% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthSim.R
\name{growthSim}
\alias{growthSim}
\title{Growth data simulating function}
\usage{
growthSim(
  model = c("logistic", "gompertz", "monomolecular", "exponential", "linear",
    "power law"),
  n = 20,
  t = 25,
  params = list(),
  noise = NULL
)
}
\arguments{
\item{df}{Data frame to use. Can be wide or long format from read.pcv}
}
\description{
growthSim can be used to help pick reasonable parameters for common growth models to use in prior distributions.

maybe I should make a data simulating function for each supported distribution as well? 
something that takes the parameters, the number of samples per group, and returns a dataframe?
}
\examples{
simdf<-growthSim("logistic", n=20, t=25, params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))
ggplot(simdf,aes(time, y, group=interaction(group,id)))+ geom_line(aes(color=group))


}
\keyword{curve,}
\keyword{exponential,}
\keyword{gompertz,}
\keyword{growth}
\keyword{law}
\keyword{linear,}
\keyword{logistic,}
\keyword{monomolecular,}
\keyword{power}
