% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{net.plot}
\alias{net.plot}
\title{Visualizing igraph networks}
\usage{
net.plot(net, fill = "strength", shape = NULL, size = 3, edgeWeight = "emd")
}
\arguments{
\item{net}{Network object similar to that returned from pcv.net, having dataframes named "edges" and "nodes"}

\item{fill}{A variable name from the nodes data to be used to color points. By default "strength" is used.}

\item{shape}{An optional discrete variable name from the nodes data to be used to change the shape of points.}

\item{size}{Size of points, defaults to 3.}

\item{edgeWeight}{Edge dataframe column to weight connections between nodes. Defaults to "emd" for compatability with \code{pcv.emd}.}

\item{...}{Further arguments passed to igraph::graph_from_dataframe or graph_from_adjacency_matrix}
}
\description{
Easy igraph visualization with pcv.net output
}
\examples{

file = "https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcvrTest1.csv"
df1<-read.pcv(file, "wide", T, multiValPattern = c("index_frequencies_index_ari", "index_frequencies_index_ci_rededge", "npq_hist_NPQ", "yii_hist_Fq'/Fm'", "yii_hist_Fv/Fm"))
colnames(df1)<-sub("index_frequencies_index_ndvi.", "ndvi_", colnames(df1))
emd_df<-pcv.emd(df1, cols="ndvi_", reorder=c("treatment", "genotype", "X"), mat =F, plot=F, parallel = 1)
network<-pcv.net(emd_df, meta = c("treatment", "genotype"))
net.plot(network, fill = "strength", shape = "genotype", size=5)

}
\keyword{distance,}
\keyword{earth}
\keyword{emd,}
\keyword{mover's}
\keyword{multi-value}
\keyword{network}
\keyword{trait,}
