% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmePlot.R
\name{nlmePlot}
\alias{nlmePlot}
\title{Function to visualize common \link{nlme::nlme} growth models.}
\usage{
nlmePlot(fit, form, df = NULL, groups = NULL, timeRange = NULL, boot = 1000)
}
\arguments{
\item{fit}{A model fit returned by \code{fitGrowth} with type="nlme".}

\item{form}{A formula similar to that in \code{growthSS} inputs (or the \code{pcvrForm} part of the output) specifying the outcome,
predictor, and grouping structure of the data as \code{outcome ~ predictor|individual/group}}

\item{df}{An optional dataframe to use in plotting observed growth curves on top of the model.
This must be supplied for nlme models.}

\item{groups}{An optional set of groups to keep in the plot.
Defaults to NULL in which case all groups in the model are plotted.}

\item{timeRange}{An optional range of times to use. This can be used to view predictions for
future data if the avaiable data has not reached some point (such as asymptotic size).}
}
\value{
Returns a ggplot showing a brms model's credible
intervals and optionally the individual growth lines.
}
\description{
Models fit using \link{growthSS} inputs by \link{fitGrowth} (and similar models made through other means)
 can be visualized easily using this function. This will generally be called by \code{growthPlot}.
}
\examples{

## Not run:

if(FALSE){

simdf<-growthSim("logistic", n=20, t=25,
params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))

ss<-growthSS(model = "logistic", form=y~time|id/group,
 df=simdf, sigma="power", start=NULL, type="nlme")
dim(ss$df)

fit<-fitGrowth(ss)

nlmePlot(fit, form = ss$pcvrForm, groups = NULL, df = ss$df, timeRange = NULL, boot=10)

}

## End(Not run)

}
\keyword{exponential,}
\keyword{gompertz,}
\keyword{growth-curve,}
\keyword{linear,}
\keyword{logistic,}
\keyword{monomolecular,}
\keyword{power-law}
