% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsPlot.R
\name{nlsPlot}
\alias{nlsPlot}
\alias{gamPlot}
\alias{lmPlot}
\title{Function to visualize common \code{stats::nls} growth models.}
\usage{
nlsPlot(
  fit,
  form,
  df = NULL,
  groups = NULL,
  timeRange = NULL,
  facetGroups = TRUE,
  groupFill = FALSE,
  virMaps = c("plasma")
)

gamPlot(
  fit,
  form,
  df = NULL,
  groups = NULL,
  timeRange = NULL,
  facetGroups = TRUE,
  groupFill = FALSE,
  virMaps = c("plasma")
)

lmPlot(
  fit,
  form,
  df = NULL,
  groups = NULL,
  timeRange = NULL,
  facetGroups = TRUE,
  groupFill = FALSE,
  virMaps = c("plasma")
)
}
\arguments{
\item{fit}{A model fit returned by \code{fitGrowth} with type="nls".}

\item{form}{A formula similar to that in \code{growthSS} inputs (or the \code{pcvrForm}
part of the output) specifying the outcome, predictor, and grouping structure of the data as
\code{outcome ~ predictor|individual/group}.
If the individual and group are specified then the observed growth lines are plotted.}

\item{df}{A dataframe to use in plotting observed growth curves on top of the model.
This must be supplied for nls models.}

\item{groups}{An optional set of groups to keep in the plot.
Defaults to NULL in which case all groups in the model are plotted.}

\item{timeRange}{An optional range of times to use. This can be used to view predictions for
future data if the avaiable data has not reached some point (such as asymptotic size).}

\item{facetGroups}{logical, should groups be separated in facets? Defaults to TRUE.}

\item{groupFill}{logical, should groups have different colors? Defaults to FALSE.
If TRUE then viridis colormaps are used in the order of virMaps}

\item{virMaps}{order of viridis maps to use. Will be recycled to necessary length.
Defaults to "plasma", but will generally be informed by growthPlot's default.}
}
\value{
Returns a ggplot showing an nls model's predictions.
}
\description{
Models fit using \link{growthSS} inputs by \link{fitGrowth}
(and similar models made through other means) can be visualized easily using this function.
This will generally be called by \code{growthPlot}.
}
\examples{

## Not run:

simdf <- growthSim("logistic",
  n = 20, t = 25,
  params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
)
ss <- growthSS(
  model = "logistic", form = y ~ time | id / group,
  df = simdf, start = NULL, type = "nls"
)
fit <- fitGrowth(ss)
nlsPlot(fit, form = ss$pcvrForm, df = ss$df, groupFill = TRUE)
nlsPlot(fit, form = ss$pcvrForm, df = ss$df, groups = "a", timeRange = 1:10)

## End(Not run)

simdf <- growthSim("logistic",
  n = 20, t = 25,
  params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
)
ss <- growthSS(
  model = "gam", form = y ~ time | id / group,
  df = simdf, start = NULL, type = "nls"
)
fit <- fitGrowth(ss)
gamPlot(fit, form = ss$pcvrForm, df = ss$df, groupFill = TRUE)
gamPlot(fit, form = ss$pcvrForm, df = ss$df, groups = "a", timeRange = 1:10)
ss <- growthSS(
  model = "gam", form = y ~ time | group,
  df = simdf, start = NULL, type = "nls"
)
fit <- fitGrowth(ss)
gamPlot(fit, form = ss$pcvrForm, df = ss$df, groupFill = TRUE)

simdf <- growthSim("logistic",
  n = 20, t = 25,
  params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
)
ss <- growthSS(
  model = "gam", form = y ~ time | id / group,
  df = simdf, start = NULL, type = "nls"
)
fit <- fitGrowth(ss)
lmPlot(fit, form = ss$pcvrForm, df = ss$df)
}
\keyword{exponential,}
\keyword{gompertz,}
\keyword{growth-curve,}
\keyword{linear,}
\keyword{logistic,}
\keyword{monomolecular,}
\keyword{power-law}
