% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvpls.R
\name{normalize}
\alias{normalize}
\title{Normalization rows or columns of a matrix}
\usage{
normalize(
  X,
  dim = 1,
  weights = if (dim == 1) 1/sqrt(rowSums(X^2)) else 1/sqrt(colSums(X^2))
)
}
\arguments{
\item{X}{matrix with numeric values}

\item{dim}{which dimension to normalize (1 for rows, 2 for columns)}

\item{weights}{vector with normalization weights, by default 2-norm is used}
}
\description{
Normalization rows or columns of a matrix
}
