% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcadf}
\alias{pcadf}
\title{Function to run a PCA, plot and optionally return the data with PCA coordinates and pca object}
\usage{
pcadf(
  df = NULL,
  cols = NULL,
  color = NULL,
  trace = NULL,
  facet = NULL,
  returnData = TRUE,
  ncp = NULL,
  split = "DAS"
)
}
\arguments{
\item{df}{Dataframe to ordinate}

\item{cols}{columns to reduce dimensions of. Can be specified with names or positions.
If this is length of 1 then it is treated as regex pattern to match
the column names that should be used.}

\item{color}{column name(s) used to color points in the pca plot.}

\item{trace}{Optional column to use to show changes over by way of geom_path,
generally this would be a time variable.}

\item{facet}{Optional column or vector to facet/split plots on.
If this is a character then it is taken as a column name.
Numeric vectors are taken as values of trace to split on,
if trace=NULL and this is numeric then plots will be split by the split argument.
 If this is a vector then a list of plots is returned.}

\item{returnData}{Logical, should data be returned?}

\item{ncp}{Optional, number of principal components to return attached
to dataframe if data is returned. Defaults to all.}

\item{split}{Time variable to split data on if a facet on time is used.
See bellwether vignette for examples.}
}
\description{
Function to run a PCA, plot and optionally return the data with PCA coordinates and pca object
}
\details{
If data is returned then it will contain the coordinates from the
 PCA and will not contain the columns that were reduced.
}
\examples{

## Not run:

dists <- list(
  rlnorm = list(meanlog = log(40), sdlog = 0.5),
  rnorm = list(mean = 60, sd = 10)
)
mv <- mvSim(dists = dists, n_samples = 100, counts = 1000,
            min_bin = 1, max_bin = 180, wide = TRUE)
pcadf(mv, cols = "sim_", returnData = TRUE)
pcadf(mv, cols = "sim_", color = c("group"), returnData = TRUE)

## End(Not run)

}
\keyword{pca}
