% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvsubread.R
\name{pcv.sub.read}
\alias{pcv.sub.read}
\title{reader function called within read.pcv when large data is used}
\usage{
pcv.sub.read(inputFile, filters, reader = "read.csv", awk = NULL, ...)
}
\arguments{
\item{inputFile}{Path to csv file of plantCV output, should be provided internally in read.pcv}

\item{filters}{filtering conditions, see read.pcv for details. Format as list("trait in area, perimeter", "other in value")}

\item{awk}{Optional awk command to use instead.}
}
\value{
Returns a dataframe after subsetting happens outside of R using the awk statement from \code{awkHelper}.
}
\description{
reader function called within read.pcv when large data is used
}
\keyword{long}
\keyword{pcv,}
\keyword{read.csv,}
\keyword{wide,}
