% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwwater.R
\name{pcv.water}
\alias{pcv.water}
\title{Read in lemnatech watering data from metadata.json files}
\usage{
pcv.water(file = NULL, envKey = "environment")
}
\arguments{
\item{file}{Path to a json file of lemnatech metadata.}

\item{envKey}{Character string representing the json key for environment data.
By default this is set to "environment".
Currently there are no situations where this makes sense to change.}
}
\value{
A data frame containing the bellwether watering data
}
\description{
Read in lemnatech watering data from metadata.json files
}
\examples{
tryCatch(
  {
    w <- pcv.water("https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/metadata.json")
  },
  error = function(e) {
    message(e)
  }
)
}
\keyword{json}
\keyword{watering}
