% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvtheme.R
\name{pcv_theme}
\alias{pcv_theme}
\title{Default theme for ggplots made by pcvr functions.}
\usage{
pcv_theme()
}
\value{
A ggplot theme
}
\description{
Default theme for ggplots made by pcvr functions.
}
\examples{
ggplot2::ggplot() +
  pcv_theme()
}
