% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvbox.R
\name{pcvBox}
\alias{pcvBox}
\title{Boxplot for plantCV data in wide format (each phenotype as a column)}
\usage{
pcvBox(
  df = df,
  x = "treatment",
  y = "area.pixels",
  fill = NULL,
  compare = F,
  trait = "trait",
  value = "value",
  ...
)
}
\arguments{
\item{df}{Data frame to use. Can be wide or long format from read.pcv}

\item{x}{X axis variable name (grouping)}

\item{y}{Y axis variable name (phenotype)}

\item{fill}{Optional variable to use for boxplot color. If set to T then the x variable is used. If this is set to a different variable than x then care should be taken with compare.}

\item{compare}{Groups to compare. By default this is set to FALSE, which corresponds to no testing. Other values of compare are passed to fixCompare to make t.test comparisons using ggpubr. In short, NULL will run all pairwise T tests, a single value of the X axis variable will compare that level to all other levels of the X variable, alternatively this can be a list as used by ggpubr: list(c("level1", "level2"), c("level1", "level3"))}

\item{trait}{If the data is in long format then `trait` is used to subset data for the phenotype Y. Defaults to "trait".}

\item{value}{If the data is in long format then `value` is used for numeric values for the phenotype Y. Defaults to "value".}

\item{...}{Additional parameters passed to geom_boxplot}
}
\description{
Boxplot for plantCV data in wide format (each phenotype as a column)
}
\examples{

df1<-read.pcv("../dataScienceEducation/troubleShootingSeminar/longPCVdata_nonElbonian.csv", "wide", T)
dim(df1)
df2<-read.pcv("../dataScienceEducation/troubleShootingSeminar/longPCVdata_nonElbonian.csv", "long", T)
dim(df2)
pcvBox(df1,x='treatment' , y='area.pixels')
pcvBox(df2, x="treatment", y="area")
pcvBox(df1,x='treatment' , y='area.pixels', fill='treatment')
}
\keyword{Bellwether,}
\keyword{ggplot}
