% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvr-package.R
\docType{package}
\name{pcvr-package}
\alias{pcvr-package}
\alias{pcvr}
\title{Plant Phenotyping and Bayesian Statistics}
\description{
\if{html}{
   \figure{pcvr_logo.png}{options: width="100" alt="logo" style="float: right"}
}

The \pkg{pcvr} package is meant to interface with any datatypes returned from \pkg{PlantCV}
including single and multi value trait phenotypes as well as common metadata for plant science
experiments as well as lower the barrier to entry for select Bayesian statistics and longitudinal
models of particular use in plant science.
}
\seealso{
Useful links:
\itemize{
\item \url{https://plantcv.org}
\item Report bugs at \url{https://github.com/danforthcenter/pcvr}
}
\code{\link{growthSS}}
\code{\link{conjugate}}
}
\author{
\strong{Maintainer}: Josh Sumner \email{jsumner@danforthcenter.org} (\href{https://orcid.org/0000-0002-3399-9063}{ORCID})

Authors:
\itemize{
  \item Jeffrey Berry (\href{https://orcid.org/0000-0002-8064-9787}{ORCID})
}

Other contributors:
\itemize{
  \item Noah Fahlgren (\href{https://orcid.org/0000-0002-5597-4537}{ORCID}) [reviewer]
  \item Donald Danforth Plant Science Center [copyright holder]
}

}
