% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrior.R
\name{plotPrior}
\alias{plotPrior}
\title{Check priors used in ease of use brms functions}
\usage{
plotPrior(priors)
}
\arguments{
\item{priors}{A named list of means for prior distributions. This does not support vectors for each element of the list or brmsprior objects, but otherwise should take the same input as \code{\link{growthSS}}.}
}
\value{
A named list of plots showing prior distributions that \code{growthSS} would make with these inputs.
}
\description{
Check priors used in ease of use brms functions
}
\examples{

priors = list("A" = 130, "B" = 10, "C" = 1.2)
priorCheck(priors)

}
\keyword{Bayesian,}
\keyword{brms}
