% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmPostPred.R
\name{postPred}
\alias{postPred}
\title{Function to visualize posterior predictive distributions from brms models
similar to those made using growthSS outputs.}
\usage{
postPred(fit, form, groups = NULL, timeRange = NULL, hyp = NULL, plot = TRUE)
}
\arguments{
\item{fit}{A brmsfit object, similar to those fit with \code{\link{growthSS}} outputs.
Alternatively this can be a list of two brmsfit objects, in which case they are compared.
If this option is used then the groups argument should be specified.}

\item{form}{A formula similar to that in \code{growthSS} inputs specifying the outcome,
predictor, and grouping structure of the data as \code{outcome ~ predictor|individual/group}.
If the fit argument is a list of two brmsfit objects then this can be a list or a single form will be
recycled to length 2.}

\item{groups}{An optional set of groups to use.
Defaults to NULL in which case the first two groups in the model are used. If the fit argument is a
list of two brmsfit objects then the first group comes from the first model and the second group
comes from the second model, in which case a NULL value here will use the first group from each
model.}

\item{timeRange}{An optional range of times to use. This can be used to view predictions for
future data if the available data has not reached some point (such as asymptotic size),
although prediction using splines outside of the observed range is not necessarily reliable.}

\item{hyp}{Optionally a character string representing a hypothesis to test. To test the difference of
two groups "diff" can be used as shorthand for group1-group2. If fit is a list then group names
should be prefixed with "fit1_" or "fit2_", respectively.}

\item{plot}{Logical, should a ggplot be returned?}
}
\value{
Returns a named list including a dataframe of draws from the posterior predictive
distribution for the specified groups, a plot if plot=TRUE and a dataframe of posterior probabilities
from brms::hypothesis if a hypothesis was specified.
}
\description{
Models fit using \link{growthSS} inputs by \link{fitGrowth}
(and similar models made through other means)
 can be visualized easily using this function. This function tests hypotheses in the original scale
 (not by growth curve parameters), visualizes the difference in predictions,
  and will generally be less powerful than brms::hypothesis used on growth model parameters.
}
\examples{

## Not run:

if (FALSE) {
  df <- growthSim("logistic",
    n = 20, t = 25,
    params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
  )
  ss <- growthSS(
    model = "logistic", form = y ~ time | id / group,
    sigma = "logistic", df = df,
    start = list(
      "A" = 130, "B" = 12, "C" = 3,
      "subA" = 20, "subB" = 10, "subC" = 1
    ), type = "brms"
  )
  fit <- fitGrowth(ss,
    iter = 1000, cores = 2, chains = 2, backend = "cmdstanr",
    control = list(adapt_delta = 0.999, max_treedepth = 20)
  )

  x <- postPred(fit, form = ss$pcvrForm, groups = NULL, timeRange = NULL, hyp = NULL, plot = TRUE)
  names(x)
  x$plot
}

## End(Not run)

}
\keyword{brms}
\keyword{growth-curve,}
