% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_class.R
\name{print.conjugate}
\alias{print.conjugate}
\title{Print a \code{conjugate} object.}
\usage{
\method{print}{conjugate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{conjugate}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Print a \code{conjugate} object.
}
\seealso{
\code{\link{summary.conjugate}}
}
