% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_class.R
\name{print.conjugatesummary}
\alias{print.conjugatesummary}
\title{Print a \code{conjugatesummary} object.}
\usage{
\method{print}{conjugatesummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{conjugatesummary}.}

\item{...}{further arguments, which are currently ignored.}
}
\description{
Print a \code{conjugatesummary} object.
}
\seealso{
\code{\link{print.conjugatesummary}}
}
