% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvrss.R
\name{print.pcvrss}
\alias{print.pcvrss}
\title{Print a \code{pcvrss} object.}
\usage{
\method{print}{pcvrss}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pcvrss}
to method \code{summary} of \code{pcvrss}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Print a \code{pcvrss} object.
}
\seealso{
\code{\link{summary.pcvrss}}
}
