% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvrss.R
\name{print.pcvrsssummary}
\alias{print.pcvrsssummary}
\title{Print a \code{pcvrsssummary} object.}
\usage{
\method{print}{pcvrsssummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pcvrsssummary}.}

\item{...}{further arguments, which are currently ignored.}
}
\description{
Print a \code{pcvrsssummary} object.
}
\seealso{
\code{\link{print.pcvrsssummary}}
}
