% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readpcvjson.R
\name{read.pcv.jsn}
\alias{read.pcv.jsn}
\title{Read in plantCV json keeping either single value, multi value, or all traits.}
\usage{
read.pcv.jsn(
  file,
  dtSuffix = "datatype",
  valSuffix = "value",
  labSuffix = "label",
  output = "sv"
)
}
\arguments{
\item{file}{Path to the plantCV output json.}

\item{dtSuffix}{Suffix for columns denoting data type. This will be one of the innermost keys in the json. Default is "datatype"}

\item{valSuffix}{Suffix for columns denoting phenotype values. This will be one of the innermost keys in the json. Default is "value"}

\item{labSuffix}{Suffix for columns denoting phenotype labels. This will be one of the innermost keys in the json. Default is "label"}

\item{output}{What phenotypes should be returned? This can be a vector of names present in "label" or one of the built in options. Built in options are "sv" for single value traits only, "mv" for multi value traits only, and "all" for all traits. Note that by default tuple value data is output with single value data in a wide format.}
}
\value{
Returns a dataframe.
}
\description{
Read in plantCV json keeping either single value, multi value, or all traits.
}
\examples{

## Not run:

file="/home/jsumner/Desktop/stargate/SINC/SINC1/plantCV_sinc1/sinc1_vis_sv_out_full.json"
x<-read.pcv.jsn(file, output="sv")
y<-read.pcv.jsn(file, output = c("area", "perimeter"))
z<-read.pcv.jsn(file, output = "mv")

## End(Not run)
}
\keyword{bellwether}
\keyword{pcv,}
\keyword{read.csv,}
