% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_helpers.R
\name{spline_diff}
\alias{spline_diff}
\title{Helper function for visualizing differences in GAMs fit with \code{mgcv::gam}}
\usage{
spline_diff(
  model,
  support,
  g1,
  g2,
  byVar,
  smoothVar,
  cis = seq(0.05, 0.95, 0.05),
  unconditional = TRUE,
  plot = TRUE
)
}
\arguments{
\item{model}{A model fit with smooth terms by \code{mgcv::gam}}

\item{support}{A data.frame of new data to use to make predictions. This should be dense support. See examples.}

\item{g1}{A character string for the level of byVar to use as the first group to compare}

\item{g2}{The second group to compare (comparison will be g1 - g2)}

\item{byVar}{Categorical variable name used to separate splines as a string.}

\item{smoothVar}{The variable that splines were used on. This will often be a time variable.}

\item{cis}{Confidence interval levels, can be multiple. For example, 0.95 would return Q_0.025 and Q_0.975 columns,
and c(0.9, 0.95) would return Q_0.025, Q_0.05, Q_0.95, and Q_0.975 columns. Defaults to \code{seq(0.05, 0.95, 0.05)}}

\item{unconditional}{Logical, should unconditional variance-covariance be used in calculating standard errors.
Defaults to TRUE.}

\item{plot}{Logical, should a plot of the difference be returned? Defaults to TRUE.}
}
\description{
Helper function for visualizing differences in GAMs fit with \code{mgcv::gam}
}
\examples{


ex <- mgcv::gamSim(eg=4,n=400,dist="normal",scale=2,verbose=FALSE)

m <- mgcv::gam(y ~ fac + s(x1, by = fac), data=ex)

support <- expand.grid(x1 = seq(min(ex$y), max(ex$y), length = 400),
                       fac = factor(levels(ex$fac)))

out<-spline_diff(m, support, "1", "2", byVar = "fac", smoothVar="x1")
dim(out$data)
out$plot

}
\keyword{GAM}
\keyword{growth,}
\keyword{mgcv,}
\keyword{spline,}
