% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ss}
\alias{ss}
\title{Example growthSS object used to make example models}
\format{
An object of class \code{pcvrss} of length 9.
}
\usage{
ss
}
\description{
A growthSS class object describing either a growth model or survival model.
}
\examples{
\donttest{
data(fit, package = "pcvr")
summary(ss)
}

}
\keyword{datasets}
