% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_brms_model.R
\docType{data}
\name{statBrmsMod}
\alias{statBrmsMod}
\title{~I don't know the export convention but it seems like this is "soft hidden"~
possible that ss is not a pcvrss object for compatibility with other brms models
if "df" is part of it then work with that otherwise use general data.
NOTE ggplot2:::Stat$compute_layer can use the default from ggproto
`make plot within a given panel of the ggplot (a facet)`
this is mostly the same as the default ggproto compute_panel function,
but it takes more named args and passes them to compute_group and
avoids warning about individual/time columns.
`make data out of model per a given aes-group, should only be 1 per panel`
this is the heavily customized component which makes data for ribbons from
the model and ss objects.}
\format{
An object of class \code{StatBrm} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
statBrmsMod
}
\description{
~I don't know the export convention but it seems like this is "soft hidden"~
possible that ss is not a pcvrss object for compatibility with other brms models
if "df" is part of it then work with that otherwise use general data.
NOTE ggplot2:::Stat$compute_layer can use the default from ggproto
`make plot within a given panel of the ggplot (a facet)`
this is mostly the same as the default ggproto compute_panel function,
but it takes more named args and passes them to compute_group and
avoids warning about individual/time columns.
`make data out of model per a given aes-group, should only be 1 per panel`
this is the heavily customized component which makes data for ribbons from
the model and ss objects.
}
\keyword{datasets}
