% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_nls_model.R
\docType{data}
\name{statNlsMod}
\alias{statNlsMod}
\title{leaving the same as statBrmsMod for now.
NOTE I think this can stay the same from brms version
`make data out of model per a given aes-group, should only be 1 per panel`
this is the heavily customized component which makes data for ribbons from
the model and ss objects.}
\format{
An object of class \code{StatNls} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
statNlsMod
}
\description{
leaving the same as statBrmsMod for now.
NOTE I think this can stay the same from brms version
`make data out of model per a given aes-group, should only be 1 per panel`
this is the heavily customized component which makes data for ribbons from
the model and ss objects.
}
\keyword{datasets}
