% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate_class.R
\name{summary.conjugate}
\alias{summary.conjugate}
\title{Summarize a \code{conjugate} object.}
\usage{
\method{summary}{conjugate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conjugate}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Summarize a \code{conjugate} object.
}
