% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvrss.R
\name{summary.pcvrss}
\alias{summary.pcvrss}
\title{Print a summary for a \code{pcvrss} object.}
\usage{
\method{summary}{pcvrss}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pcvrss}.}

\item{...}{further arguments, which are currently ignored.}
}
\description{
Print a summary for a \code{pcvrss} object.
}
\seealso{
\code{\link{print.pcvrss}}
}
