% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{surv}
\alias{surv}
\title{Example survival model}
\format{
A brmsfit object and a growthss object
}
\usage{
surv
}
\description{
An exponential survival brms model of simulated data.
}
\examples{
\donttest{
data(surv, package = "pcvr")
summary(surv)
}

}
\keyword{datasets}
