% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellwethergrowth.R
\name{bellwetherGrowth}
\alias{bellwetherGrowth}
\title{Growth curve plot for bellwether data}
\usage{
bellwetherGrowth(
  df = NULL,
  x = "DAS",
  y = "area_adj",
  group = c("Barcodes", "angle"),
  warn = T,
  color = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame to use. Should be in wide format similar to output from read.pcv.bw}

\item{x}{X axis variable name (time variable)}

\item{y}{Y axis variable name (phenotype)}

\item{group}{Set of variables to use in grouping observations. These taken together should identify a unique plant (or unique plant at a unique angle) across time.}

\item{warn}{Logical, should a warning be printed if there are duplicate values for a timepoint (`x`) using the `group` variables. If TRUE then the warning will suggest how to find duplicates to amend `group`.}

\item{color}{Optional variable to use for line colors.}

\item{...}{Additional parameters passed to geom_line}
}
\description{
Growth curve plot for bellwether data
}
\examples{
bw<-read.pcv.bw( file="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestPhenos.csv", snapshotFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestSnapshot.csv", designFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestDesign.csv",metaCol="meta",metaForm="vis_view_angle_zoom_horizontal_gain_exposure_v_new_n_rep",joinSnapshot="id",conversions = list(area=13.2*3.7/46856) )
bellwetherGrowth(df=bw,x='DAS' , y='area_adj', group=c('Barcodes',"angle") )

}
\keyword{Bellwether,}
\keyword{ggplot}
