% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmPlot.R
\name{brmPlot}
\alias{brmPlot}
\title{Function to visualize brms models similar to those made using growthSS outputs.}
\usage{
brmPlot(fit, form, df = NULL)
}
\arguments{
\item{fit}{A brmsfit object, similar to those fit with \code{\link{growthSS}} outputs.}

\item{form}{A formula similar to that in \code{growthSS} inputs specifying the outcome, predictor, and grouping structure of the data as \code{outcome ~ predictor|individual/group}.}

\item{df}{An optional dataframe to use in plotting observed growth curves on top of the model.}
}
\description{
Function to visualize brms models similar to those made using growthSS outputs.
}
\examples{
simdf<-growthSim("logistic", n=20, t=25, params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))
ss<-growthSS(model = "logistic", form=y~time|id/group, sigma="spline", df=simdf, priors = list("A"=130, "B"=12, "C"=3))
lapply(ss,class)
ss$initfun()
fit_test <- brm(ss$formula, prior = ss$prior, data = ss$df, family = ss$family, # main componenets of the model
              iter = 1000, cores = 2, chains = 2, init = ss$initfun, # parameters controling chain number, chain length, parallelization and starting values
              control = list(adapt_delta = 0.999, max_treedepth = 20), backend = "cmdstanr") # options to increase performance
brmPlot(fit_test, y~time|id/group, df=NULL)
print(load(url("https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/brmsFits.rdata")))
brmPlot(fit_25, form = y~time|id/group)
brmPlot(fit_9, form = y~time|id/group)
brmPlot(fit_15, form = y~time|id/group)

}
\keyword{curve,}
\keyword{exponential,}
\keyword{gompertz,}
\keyword{growth}
\keyword{law}
\keyword{linear,}
\keyword{logistic,}
\keyword{monomolecular,}
\keyword{power}
