% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmPlot.R
\name{brmPlot}
\alias{brmPlot}
\title{Function to visualize brms models similar to those made using growthSS outputs.}
\usage{
brmPlot(fit, form, groups = NULL, df = NULL, timeRange = NULL)
}
\arguments{
\item{fit}{A brmsfit object, similar to those fit with \code{\link{growthSS}} outputs.}

\item{form}{A formula similar to that in \code{growthSS} inputs specifying the outcome,
predictor, and grouping structure of the data as \code{outcome ~ predictor|individual/group}.}

\item{groups}{An optional set of groups to keep in the plot.
Defaults to NULL in which case all groups in the model are plotted.}

\item{df}{An optional dataframe to use in plotting observed growth curves on top of the model.}

\item{timeRange}{An optional range of times to use. This can be used to view predictions for
future data if the avaiable data has not reached some point (such as asymptotic size),
although prediction using splines outside of the observed range is not necessarily reliable.}
}
\value{
Returns a ggplot showing a brms model's credible
intervals and optionally the individual growth lines.
}
\description{
Function to visualize brms models similar to those made using growthSS outputs.
}
\examples{

## Not run:

if(FALSE){

data(bw_vignette_fit)
brmPlot(bw_vignette_fit, y~time|id/group, df=NULL)
print(load(url("https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/brmsFits.rdata")))
brmPlot(fit_25, form = y~time|id/group)
brmPlot(fit_9, form = y~time|id/group)
brmPlot(fit_15, form = y~time|id/group)

}

## End(Not run)

}
\keyword{exponential,}
\keyword{gompertz,}
\keyword{growth-curve,}
\keyword{linear,}
\keyword{logistic,}
\keyword{monomolecular,}
\keyword{power-law}
