% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwoutliers.R
\name{bw.outliers}
\alias{bw.outliers}
\title{Remove outliers from bellwether data using cook's distance}
\usage{
bw.outliers(
  df = NULL,
  phenotype = "area_adj",
  naTo0 = F,
  group = c("DAS", "Genotype"),
  plotgroup = c("Barcodes", "angle"),
  plot = T,
  x = NULL
)
}
\arguments{
\item{df}{Data frame to use. Should be in the format of output from read.pcv.bw}

\item{phenotype}{Column to use to classify outliers.}

\item{naTo0}{Logical, should NA values to changed to 0.}

\item{group}{Grouping variables to find outliers as a character vector. This is typically time  and design variables (DAS, genotype, treatment, etc). These are used as predictors for `phenotype` in a generalized linear model.}

\item{plot}{Logical, if TRUE then a list is returned with a ggplot and a dataframe.}

\item{x}{Optional specification for x axis variable if plot is true. If left NULL (the default) then the first element of `group` is used.}
}
\description{
Remove outliers from bellwether data using cook's distance
}
\examples{
bw<-read.pcv.bw( file="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestPhenos.csv", snapshotFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestSnapshot.csv", designFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestDesign.csv",metaCol="meta",metaForm="vis_view_angle_zoom_horizontal_gain_exposure_v_new_n_rep",joinSnapshot="id",conversions = list(area=13.2*3.7/46856) )
bw<-bw.outliers(df = bw, phenotype="area_adj", naTo0 = F, group = c("DAS", "Genotype"), plotgroup=c('Barcodes',"angle"), plot=T)
}
\keyword{Bellwether,}
\keyword{ggplot}
