% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwtime.R
\name{bw.time}
\alias{bw.time}
\title{Time conversion and plotting for bellwether data}
\usage{
bw.time(
  df = NULL,
  mode = NULL,
  plantingDelay = NULL,
  phenotype = NULL,
  cutoff = 1,
  timeCol = "timestamp",
  group = "Barcodes",
  plot = T,
  format = "\%Y-\%m-\%d \%H:\%M:\%S"
)
}
\arguments{
\item{df}{Data frame to use. Should be in the format of output from read.pcv.bw}

\item{mode}{One of "DAS", "DAP" or "DAE" (Days After Planting and Days After Emergence). Defaults to NULL in which case all columns are added. Note that if timeCol is not an integer then DAS is always returned.}

\item{plantingDelay}{If `mode` includes "DAP" then `plantingDelay` is used to adjust "DAS"}

\item{phenotype}{If `mode` includes "DAE" then this is the phenotype used to classify emergence.}

\item{cutoff}{If `mode` inlcludes "DAE" then this value is used to classify emergence. Defaults to 1, meaning an image with a value of 1 or more for `phenotype` has "emerged".}

\item{timeCol}{Column of input time values, defaults to "timestamp". If this is not an integer then it is assumed to be a timestamp in the format of the format argument.}

\item{group}{Grouping variables to specify unique plants as a character vector. This defaults to "Barcodes". These taken together should identify a unique plant across time, although often "angle" or "rotation" should be added.}

\item{plot}{Logical, should plots of the new time variables be printed?}

\item{format}{An R POSIXct format, defaults to lemnatech standard format. This is only used if timeCol is not an integer.}
}
\value{
The input dataframe with new integer columns for different ways of describing time in the experiment.
}
\description{
Time conversion and plotting for bellwether data
}
\examples{
bw<-read.pcv.bw( file="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestPhenos.csv", snapshotFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestSnapshot.csv", designFile="https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/bwTestDesign.csv",metaCol="meta",metaForm="vis_view_angle_zoom_horizontal_gain_exposure_v_new_n_rep",joinSnapshot="id",conversions = list(area=13.2*3.7/46856) )
bw<-bw.time(df = bw, mode=NULL, plantingDelay = 4, phenotype="area_adj", timeCol="DAS", group="Barcodes", plot=T)

}
\keyword{Bellwether,}
\keyword{ggplot}
