% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGrowth.R
\name{fitGrowth}
\alias{fitGrowth}
\title{Ease of use wrapper function for fitting various growth models specified by \code{growthSS}}
\usage{
fitGrowth(ss, ...)
}
\arguments{
\item{ss}{A list generated by \code{growthSS}.}

\item{...}{Additional arguments passed to model fitting functions determined by \code{ss$type}. 
For type = "nlme" these are passed to \code{nlme::nlmeControl}, not \code{nlme::nlme}.
Additional arguments are documented in \link{fitGrowthBrm}, \link{fitGrowthNlme}, \link{fitGrowthNls}, \link{fitGrowthNlrq},
\link{fitGrowthMgcv}, \link{fitGrowthSurvReg}, \link{fitGrowthFlexSurvReg}.}
}
\value{
A fit model from the selected type.
}
\description{
Ease of use wrapper function for fitting various growth models specified by \code{growthSS}
}
\examples{

## Not run:
simdf<-growthSim("logistic", n=20, t=25,
   params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))
ss<-growthSS(model = "logistic", form=y~time|id/group, sigma="spline",
  df=simdf, start = list("A"=130, "B"=12, "C"=3), type="brms")
lapply(ss,class)
ss$initfun()

if(FALSE){
fit_test <- fitGrowth(ss, iter = 600, cores = 2, chains = 2, backend = "cmdstanr",
              control = list(adapt_delta = 0.999, max_treedepth = 20)) # optional controls
}
## End(Not run)
}
\keyword{Bayesian,}
\keyword{brms,}
\keyword{nlme,}
\keyword{nlrq}
\keyword{nls,}
