% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvSS.R
\name{mvSS}
\alias{mvSS}
\title{Ease of use multi-value trait model helper function.}
\usage{
mvSS()
}
\arguments{
\item{form}{A formula similar to \code{label | value ~ time + id/group} where label is a column
of histogram bins, value is the counts within those bins, time is an optional time variable,
id identifies an individual, and group contains the treatment groups.}

\item{sigma}{Distributional models passed to \link{growthSS}.}

\item{df}{Data passed to \link{growthSS}.}

\item{pars}{Parameters to vary, passed to \link{growthSS}.}

\item{start}{Starting values or priors, passed to \link{growthSS}.}

\item{type}{Backend to use, passed to \link{growthSS}.}

\item{tau}{Quantile to model, passed to \link{growthSS}.}

\item{hierarchy}{Formulae describing any hierarchical models, see \link{growthSS}.}

\item{spectral_index}{A spectral index
\href{https://plantcv.readthedocs.io/en/stable/spectral_index/}{from those calculated by PlantCV}}
}
\value{
A named list of plots showing prior distributions that \code{growthSS} would use,
optionally with a plot of simulated growth curves using draws from those priors.
}
\description{
This function provides a simplified interface to modeling multi-value traits using \link{growthSS}.
Output from this should be passed to \link{fitGrowth} to fit the specified model.
}
\examples{
mv_df <- mvSim(dists = list(rnorm = list(mean = 100, sd = 30)), wide = FALSE)
mv_df <- mv_df[mv_df$value > 0, ]
# show that this has identical output to a more complicated call to growthSS

}
\keyword{multi-value}
