% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netplot.R
\name{net.plot}
\alias{net.plot}
\title{Visualizing igraph networks}
\usage{
net.plot(
  net,
  fill = "strength",
  shape = NULL,
  size = 3,
  edgeWeight = "emd",
  edgeFilter = NULL
)
}
\arguments{
\item{net}{Network object similar to that returned from pcv.net, having dataframes named "edges" and "nodes"}

\item{fill}{Variable name(s) from the nodes data to be used to color points. By default "strength" is used.}

\item{shape}{Optional discrete variable name(s) from the nodes data to be used to change the shape of points.
If this variable is numeric it will be coerced to character.}

\item{size}{Size of points, defaults to 3.}

\item{edgeWeight}{Edge dataframe column to weight connections between nodes. Defaults to "emd" 
for compatability with \code{pcv.emd}.}

\item{edgeFilter}{How should edges be filtered? This can be either a numeric (0.5) 
in which case it is taken as a filter where only edges with values greater than or equal to 
that number are kept or a character string ("0.5") in which case the strongest 
X percentage of edges are kept. This defaults to NULL which does no filtering, 
although that should not be considered the best standard behaviour. See details.}
}
\value{
Returns a ggplot of a network.
}
\description{
Easy igraph visualization with pcv.net output
}
\examples{

## Not run:

file = "https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcvrTest1.csv"
df1<-read.pcv(file, "wide", TRUE, multiValPattern = c("index_frequencies_index_ari",
 "index_frequencies_index_ci_rededge", "npq_hist_NPQ", "yii_hist_Fq'/Fm'", "yii_hist_Fv/Fm"))
 
colnames(df1)<-sub("index_frequencies_index_ndvi.", "ndvi_", colnames(df1))

w<-pcv.emd(df1, cols="ndvi_", reorder=c("treatment", "genotype"), 
  mat =FALSE, plot=FALSE, parallel = 1)
  
network<-pcv.net(w, meta = c("treatment", "genotype"))
net=network; fill = "strength"; shape = "genotype"; size=5; edgeWeight="emd"
net.plot(network, fill = "strength", shape = "genotype", size=5, edgeFilter=0.5)
net.plot(network, fill = "strength", shape = "genotype", size=5, edgeFilter="0.5")

## End(Not run)

}
\keyword{distance,}
\keyword{earth}
\keyword{emd,}
\keyword{mover's}
\keyword{multi-value}
\keyword{network}
\keyword{trait,}
