% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvjoyplot.R
\name{pcv.joyplot}
\alias{pcv.joyplot}
\title{Make Joyplots for multi value trait plantCV data and optionally compare distributions}
\usage{
pcv.joyplot(
  df = NULL,
  index = NULL,
  group = NULL,
  method = NULL,
  compare = F,
  priors = NULL,
  hyp = NULL,
  bin = "label",
  freq = "value",
  trait = "trait"
)
}
\arguments{
\item{df}{Data frame to use. Long format with multi value traits is expected.}

\item{index}{A multi value trait as a character string. Must be present in `trait`.}

\item{group}{A length 1 or 2 character vector. This is used for faceting the joyplot and identifying groups for testing.}

\item{method}{A method to use in comparing distributions/means. Currently "beta", "gaussian", "ks", and "emd" are supported. See details for explanations of tests.}

\item{compare}{Groups to compare. By default this is set to FALSE, which corresponds to no testing. Other values of compare are passed to fixCompare to make t.test comparisons using ggpubr. In short, NULL will run all pairwise T tests, a single value of the X axis variable will compare that level to all other levels of the X variable, alternatively this can be a list as used by ggpubr: list(c("level1", "level2"), c("level1", "level3"))}

\item{priors}{Parameters for prior distributions if using method = "beta" or "gaussian". If left NULL then wide weak priors are used. This can be set as a single set of parameters or with unique values for each group in `compare`, but it is advised to use the default priors.}

\item{hyp}{Hypothesis to test for beta and gaussian methods, one of "unequal", "greater", "lesser". Defaults to "unequal" if left NULL.}

\item{bin}{Column containing histogram (multi value trait) bins. Defaults to "label".}

\item{freq}{Column containing histogram counts. Defaults to "value"}

\item{trait}{Column containing phenotype names. Defaults to "trait".}
}
\description{
Make Joyplots for multi value trait plantCV data and optionally compare distributions
}
\details{
The method argument is used for statistical testing of groups. There are currently four options, "beta", "gaussian", "ks", and "emd".
\itemize{
 \item{"beta"}{The beta method is for use with bounded data (0,1) and creates a posterior beta distribution using the histogram counts and bins and priors (which default to beta(0.5,0.5)). The posterior distributions are compared across the groups in `compare` and a posterior probability is returned for each which corresponds to P[hyp]. }
 \item{"gaussian"}{The gaussian method is for use with multi value traits that are not bounded 0-1 and are roughly gaussian. Like the beta method this assumes weak priors and uses the histogram data to make a posterior distribution (here a T distribution). Here there will be two posterior probabilities returned for each comparison, one compares the distribution of means only and is roughly analogous to a T test, the other compares the entire posterior distributions and is more similar to a Z test. The posterior probability still takes the form P[hyp]. Note that currently these consider pixels as reps, so it is easy to have very small biological changes at very high posterior probability.}
 \item{"ks"}{The ks method performs a ks test on the PDF of each histogram over a set support range. This returns a P value corresponding to a standard KS test that distributions are the same. For this method `hyp` is ignored.}
 \item{"emd}{The emd method returns earth mover's distance for compared histograms. This is done via the emdist package and can be very slow. This is included because it is canonical but in the case where EMD is desired it may be worth implementing it directly.}
}
}
\examples{
df <- read.pcv("https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcvrTest2.csv", "long", F)
pcv.joyplot(df, index = "index_frequencies_index_ndvi", group=c("genotype", "timepoint"))
pcv.joyplot(df, index = "index_frequencies_index_ndvi", group=c("genotype", "timepoint"), method="ks")
pcv.joyplot(df, index = "index_frequencies_index_ndvi", group=c("genotype", "timepoint"), method="beta")
pcv.joyplot(df, index = "index_frequencies_index_ndvi", group=c("genotype", "timepoint"), method="gaussian")
}
\keyword{bayesian,}
\keyword{ggplot,}
\keyword{multi}
\keyword{pcv.hists}
\keyword{trait,}
\keyword{value}
