% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{pcv.net}
\alias{pcv.net}
\title{Network analysis of a distance matrix}
\usage{
pcv.net(
  emd = NULL,
  meta = NULL,
  dissim = T,
  distCol = "emd",
  filter = NULL,
  direction = "greater"
)
}
\arguments{
\item{emd}{A long dataframe as returned by pcv.emd. Currently this function is only made to work with dataframe output, not distance matrix output.}

\item{meta}{Metadata to be carried from pcv.emd output into the network, defaults to NULL which will use all metadata. Type conversion will be attempted for these columns.}

\item{dissim}{Logical, should the distCol be inverted to make a dissimilarity value?}

\item{distCol}{The name of the column containing distances/dissimilarities. Defaults to "emd" for compatability with pcv.emd}

\item{filter}{This can be either a numeric (0.5) in which case it is taken as a filter where only edges with values greater than or equal to that number are kept or a character string ("0.5") in which case the strongest X percentage of edges are kept.
This defaults to NULL which does no filtering, although that should not be considered the best standard behavior. Note that this will happen after converting to dissimilarity if dissim=T.}

\item{direction}{Direction of filtering, can be either "greater" or "lesser".}
}
\value{
Returns a list containing three elements:
\code{nodes}: A dataframe of node data.
\code{edges}: A dataframe of edges between nodes.
\code{graph}: The network as an igraph object
}
\description{
Easy igraph use with pcv.emd output
}
\examples{

file = "https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcvrTest1.csv"
df1<-read.pcv(file, "wide", T, multiValPattern = c("index_frequencies_index_ari", "index_frequencies_index_ci_rededge", "npq_hist_NPQ", "yii_hist_Fq'/Fm'", "yii_hist_Fv/Fm"))
colnames(df1)<-sub("index_frequencies_index_ndvi.", "ndvi_", colnames(df1))
emd_df<-pcv.emd(df1, cols="ndvi_", reorder=c("treatment", "genotype"), mat =F, plot=F, parallel = 1)
net<-pcv.net(emd_df, meta = c("treatment", "genotype"))

}
\keyword{distance,}
\keyword{earth}
\keyword{emd,}
\keyword{mover's}
\keyword{multi-value}
\keyword{network}
\keyword{trait,}
