% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsr.R
\name{pcv.plsr}
\alias{pcv.plsr}
\title{Run Partial Least Squares Regression on spectral data}
\usage{
pcv.plsr(df, resps = NULL, spectra = NULL, train = 0.8, cv = 10, ...)
}
\arguments{
\item{df}{Data frame containing metadata and spectral histogram data}

\item{resps}{Vector of response variables.}

\item{spectra}{Either one column name (in the case of long data) or a set of columns in the case of wide data.
If a single character string is provided and it is not one of the column names then it is
taken to be a pattern that will match some set of column names in the data to use (see examples).}

\item{train}{Proportion of data to use as training data.}

\item{cv}{Number of cross validation iterations.}

\item{...}{Further arguments passed to caret::train.}
}
\description{
Partial Least Squares Regression (plsr) is often used to analyze spectral data.
}
\details{
Note that columns that sum to 0 in the training or test data will be removed.
}
\examples{

## Not run:

file = "https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcvrTest1.csv"
df1<-read.pcv(file, "wide", T, multiValPattern = c("index_frequencies_index_ari",
"index_frequencies_index_ci_rededge", "npq_hist_NPQ", "yii_hist_Fq'/Fm'", "yii_hist_Fv/Fm"))
colnames(df1)<-sub("index_frequencies_index_ndvi.", "ndvi_", colnames(df1))
x<-pcv.plsr(df=df1, resps = "area.pixels", spectra = grepl("^ndvi_", colnames(df1)))
x<-pcv.plsr(df=df1, resps = "area.pixels", spectra = "^ndvi_")

## End(Not run)

}
\keyword{PLSR}
