% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wue.R
\name{pwue}
\alias{pwue}
\title{Calculate pseudo water use efficiency from phenotype and watering data}
\usage{
pwue(df, w, pheno = "area", time = "DAS", id = "barcode")
}
\arguments{
\item{df}{Dataframe containing wide single-value phenotype data.
This should already be aggregated to one row per plant per day (angles/rotations combined).}

\item{w}{Watering data as returned from bw.water.}

\item{pheno}{Phenotype column name, defaults to "area"}

\item{time}{Variable(s) that identify a plant on a given day.
Defaults to \code{c("barcode", "DAS")}.}

\item{id}{Variable(s) that identify a plant over time. Defaults to \code{"barcode"}.}
}
\value{
A data frame containing the bellwether watering data joined
    to phenotype data with new columns for change in the phenotype,
    change in the pre-watering weight, and pseudo-water use efficiency (pWUE).
}
\description{
Water use efficiency (WUE) is the change in biomass per unit of water metabolized.
Using image based phenotypes and watering data we can calculate pseudo-WUE (pwue) over time.
Here area is used as a proxy for biomass and transpiration is approximated using watering data.
The equation is then \eqn{\frac{P_{t} - P_{t-1}}{W_{t_{end-1}}-W_{t_{start}} }}{P_[t] - P_[t-1] / W_[t_(end-1)]-W_[t_start]},
where P is the phenotype and W is the weight before watering.
}
\examples{

## Not run:

library(data.table)
sv<-read.pcv(
   "https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/pcv4-single-value-traits.csv",
   reader="fread")
sv$genotype = substr(sv$barcode, 3,5)
sv$genotype = ifelse(sv$genotype == "002", "B73",
     ifelse(sv$genotype == "003", "W605S",
     ifelse(sv$genotype == "004", "MM", "Mo17")))
sv$fertilizer = substr(sv$barcode, 8, 8)
sv$fertilizer = ifelse(sv$fertilizer == "A", "100",
           ifelse(sv$fertilizer == "B", "50", "0"))
sv<-bw.time(sv, plantingDelay = 0,
      phenotype="area", cutoff=10, timeCol="timestamp",
      group=c("barcode", "rotation"), plot=TRUE)
phenotypes <- colnames(sv)[19:35]
phenoForm<-paste0("cbind(", paste0(phenotypes, collapse=", "), ")")
groupForm<-"DAS+DAP+barcode+genotype+fertilizer"
sv<-aggregate(as.formula(paste0(phenoForm, "~", groupForm)), data=sv, mean, na.rm=TRUE)
sv<-bw.outliers(sv, phenotype="area",
      group = c("DAS", "genotype", "fertilizer"), plotgroup = c("barcode"))
water<-bw.water("https://raw.githubusercontent.com/joshqsumner/pcvrTestData/main/metadata.json")

df<-sv
w<-water
pheno="area"
id =c("barcode")
time="DAS"

x<-pwue(df, w, pheno, time, id)
library(ggplot2)
ggplot(x, aes(x=DAS, y=pWUE, group = barcode))+
  geom_line()+
  theme_minimal()
 
head(x[which(x$pWUE< -100),])
head(x[x$pWUE < -100,]) # note the counter-intuitive handling of NAs in the index column.
# Either use data.table or wrap conditions in which().

## End(Not run)


}
\keyword{long}
\keyword{pcv,}
\keyword{read.csv,}
\keyword{wide,}
