% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_brms_model.R, R/stat_growthSS.R,
%   R/stat_nlme_model.R, R/stat_nlrq_model.R, R/stat_nls_model.R
\name{stat_brms_model}
\alias{stat_brms_model}
\alias{stat_growthss}
\alias{stat_nlme_model}
\alias{stat_nlrq_model}
\alias{stat_nls_model}
\title{Show brms model in ggplot layer}
\usage{
stat_brms_model(
  mapping = NULL,
  data = NULL,
  fit = NULL,
  ss = NULL,
  CI = 0.95,
  inherit.aes = TRUE,
  ...
)

stat_growthss(
  mapping = NULL,
  data = NULL,
  fit = NULL,
  ss = NULL,
  inherit.aes = TRUE,
  ...
)

stat_nlme_model(
  mapping = NULL,
  data = NULL,
  fit = NULL,
  ss = NULL,
  inherit.aes = TRUE,
  ...
)

stat_nlrq_model(
  mapping = NULL,
  data = NULL,
  fit = NULL,
  ss = NULL,
  inherit.aes = TRUE,
  ...
)

stat_nls_model(
  mapping = NULL,
  data = NULL,
  fit = NULL,
  ss = NULL,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{ggplot2::aes()}.
If specified and ‘inherit.aes = TRUE’ (the default),
it is combined with the default mapping at the top level of the plot.
If there is no mapping then it is filled in by default using the \code{ss} object.}

\item{data}{The data to be displayed in this layer. This behaves per normal ggplot2 expectations
except that if data is missing (ie, not inherited or specified) then the data from \code{ss} is used.}

\item{fit}{A brmsfit object, typically returned from \code{fitGrowth}.}

\item{ss}{A \code{pcvrss} object. Only the "pcvrForm" and "df" elements are used.}

\item{CI}{A vector of credible intervals to plot, defaults to 0.95.}

\item{inherit.aes}{Logical, should aesthetics be inherited from top level? Defaults to TRUE.}

\item{...}{Additional arguments passed to ggplot2::layer.}
}
\description{
Add posterior predictive distribution from a brms model fit with \link{growthSS} and \link{fitGrowth}
to a ggplot object.

Add a model fit with \link{growthSS} and \link{fitGrowth} to a ggplot object.
The exact geom used depends on the model class (see details).

Add predicted mean trendline from a nlme/lme model fit with \link{growthSS} and \link{fitGrowth}
to a ggplot object.

Add predicted mean trendline from a model fit with \link{growthSS} and \link{fitGrowth}
to a ggplot object.

Add predicted mean trendline from a model fit with \link{growthSS} and \link{fitGrowth}
to a ggplot object.
}
\details{
These layers will behave largely like output from \code{\link{growthPlot}}, although \code{growthPlot}
has more arguments that directly control the plot since this stat only makes one layer.
The geometries used for each type of model are:
\itemize{
  \item{\strong{brms}: \code{geom_ribbon} for longitudinal plots, \code{geom_rect} for others.}
  \item{\strong{nlrq}: \code{geom_line}, replicated per each quantile.}
  \item{\strong{nlme}: \code{geom_smooth}, with ribbon based on the heteroskedastic term.}
  \item{\strong{nls}: \code{geom_line}, replicated per each quantile.}
  \item{\strong{nlrq}: \code{geom_line}, replicated per each quantile.}
}
}
\examples{
library(ggplot2)
simdf <- growthSim("logistic",
  n = 20, t = 25,
  params = list("A" = c(200, 160), "B" = c(13, 11), "C" = c(3, 3.5))
)
ss <- growthSS(
  model = "logistic", form = y ~ time | id / group,
  df = simdf, start = NULL, type = "nls"
)
fit <- fitGrowth(ss)
ggplot() +
  stat_growthss(fit = fit, ss = ss)

}
\seealso{
\link{growthPlot} for a self-contained plotting function
}
\keyword{ggplot}
