% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcvrss.R
\name{summary.pcvrss}
\alias{summary.pcvrss}
\title{Summarize a \code{pcvrss} object.}
\usage{
\method{summary}{pcvrss}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pcvrss}
to method \code{summary} of \code{pcvrss}.}

\item{...}{further arguments, passed to print.default.}
}
\description{
Summarize a \code{pcvrss} object.
}
