% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGrowth.R
\name{testGrowth}
\alias{testGrowth}
\title{Hypothesis testing for frequentist \code{fitGrowth} models.}
\usage{
testGrowth(ss, fit, test_pars = "A")
}
\arguments{
\item{ss}{A list output from \link{growthSS}.}

\item{fit}{A non-brms model (or list of nlrq models) output from \link{fitGrowth}.}

\item{test_pars}{A vector of parameters to test in the model. These should be parameters which
vary by group in your original model and that you want to test against a null model where they
do not vary by group. Alternatively for nlrq models this can be a comparison of model terms
written as \code{"group_X|tau|par > group_Y|tau|par"}, which uses a fat tailed T distribution to make 
comparisons on the means of each quantile estimate.}
}
\value{
A list containing an anova object comparing non-linear growth models and the null model.
}
\description{
Hypothesis testing for frequentist \code{fitGrowth} models.
}
\details{
For nls and nlme models an anova is run and returned as part of a list along with the null model.
 For nlrq models several assumptions are made and a likelihood ratio test for each tau 
 is run and returned as a list.
}
\examples{

## Not run:
set.seed(123)
simdf<-growthSim("logistic", n=20, t=25,
   params = list("A"=c(200,160), "B"=c(13, 11), "C"=c(3, 3.5)))
ss<-suppressMessages(growthSS(model = "logistic", form=y~time|id/group,
  df=simdf, type="nlrq"))
fit <- fitGrowth(ss)
testGrowth(ss, fit, "A")

## End(Not run)
}
\keyword{hypothesis,}
\keyword{nlme,}
\keyword{nlrq}
\keyword{nls,}
